refactor: move transactions, auth and user to modules folder
