+++ ms-users/src/modules/users/user.model.ts
@@ -1,5 +1,5 @@
 import { DataTypes, Model, Optional } from 'sequelize';
-import { sequelize } from '../config/database';
+import { sequelize } from '../../config/database';
 
 export interface UserAttributes {
   id: string;
