+++ ms-wallet/src/modules/transactions/transaction.model.ts
@@ -1,5 +1,5 @@
 import { DataTypes, Model, Optional } from "sequelize";
-import { sequelize } from "../config/database";
+import { sequelize } from "../../config/database";
 
 export enum TransactionType {
   CREDIT = 'CREDIT',
